complete userController and authService
Complete two api functions in userController: getProfile and
updateProfile. Complete the authenticate api functions:
regist, login, logout. The prototype for OAuth is in function
googleAuth.

Note that we should use authMiddleware to checkt the token.
Besides, we can use result[row][item name] to get the element
in the sql query result.