update userController and test script
In userController, I remove the middleware module because it would
do in the app.js. In the authService, we can remove it similarly.

In the user.test.js file, when it get tokena, it will not assign
the id in it to uida, so that the expected output is undefined.
Use jwt verification function to get it.

Modify package.json so that jest should run sequentially, to
avoid the race condition in shared database.